+++ CrypTool/IntegerArithmetic.cpp
@@ -541,7 +541,7 @@ int decode( const char *StrNumber, char *data, size_t datalen, int blockLength, 
 	   	t = t / modul;
        	tmp[sizeof(tmp)-(i+1)] = ch;
 	}
-	ASSERT(i + 1 < datalen);
+	ASSERT((unsigned int)(i + 1) < datalen);
 
 	int j;
 	for (j=sizeof(tmp)-i; j<sizeof(tmp); j++ )
